+++ test/test-buildDevLoaderFile.js
@@ -44,7 +44,6 @@ describe(name, async () => {
 			
 			assert.deepEqual(created, [
 				'dev/site.css',
-				'mojl_cache/hashes/src/c/icon.gif.mojlcache',
 			]);
 			assert(after["dev/site.css"].includes(expectedCSS.trim()));
 
@@ -100,7 +99,6 @@ describe(name, async () => {
 			assert.deepEqual(created, [
 				'dev/site.css',
 				'dev/site.js',
-				'mojl_cache/hashes/src/c/icon.gif.mojlcache',
 			]);
 			assert(after["dev/site.css"].includes(expectedCSS.trim()));
 			assert(after["dev/site.js"].includes(expectedJS.trim()));